chore: renaming and flattening navigation to: Navigation, AuthNavigation and HomeNavigation, added createFixture with Rootstate fixture as default
