+++ .github/pull_request_template.md
@@ -1,7 +1,6 @@
-![](gifSrc)
 ## Scope [YOMA-XXX](https://yomaworld.atlassian.net/browse/YOMA-XXX)
 
-
+![](gifSrc)
 ## Work Done
