--- src/modules/AppNavigation/Authentication/Authentication.routes.ts
@@ -1,8 +1,0 @@
-export enum AuthNavigationRoutes {
-  Landing = 'Landing',
-  Register = 'Register',
-  RegisterWithEmail = 'RegisterWithEmail',
-  Login = 'Login',
-  ForgotPassword = 'ForgotPassword',
-  ResetPassword = 'ResetPassword',
-}
