--- src/modules/AppNavigation/index.ts
@@ -1,3 +1,0 @@
-import AppNavigation from './AppNavigation.container'
-
-export default AppNavigation
