+++ src/modules/Auth/Auth.selector.ts
@@ -4,5 +4,5 @@ import { RootState } from 'redux/redux.types'
 
 export const selectAuthState = (state: RootState) => state.auth
 export const selectToken = createSelector(selectAuthState, propOr(null, 'token'))
-export const selectIsAuthenticated = createSelector(selectToken, complement(isNil))
+export const selectIsAuthenticated: (state: RootState) => boolean = createSelector(selectToken, complement(isNil))
 export const selectLoginCredentials = pipe(path(['auth']), pick(['email', 'password']))
