+++ src/modules/AuthNavigation/AuthNavigation.types.ts
@@ -1,4 +1,11 @@
-import { AuthNavigationRoutes } from './Authentication.routes'
+export enum AuthNavigationRoutes {
+  Landing = 'Landing',
+  Register = 'Register',
+  RegisterWithEmail = 'RegisterWithEmail',
+  Login = 'Login',
+  ForgotPassword = 'ForgotPassword',
+  ResetPassword = 'ResetPassword',
+}
 
 export type AuthNavigatorParamsList = {
   [AuthNavigationRoutes.Landing]: undefined
