+++ src/modules/DigitalCv/DigitalCv.tsx
@@ -1,7 +1,6 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { ViewContainer, HomeHeader, FirstTimeCard, CvCard } from 'components'
-import { HomeNavigationRoutes } from 'modules/Home/Home.routes'
-import { HomeNavigatorParamsList } from 'modules/Home/Home.types'
+import { HomeNavigationRoutes, HomeNavigatorParamsList } from 'modules/HomeNavigation/HomeNavigation.types'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ScrollView } from 'react-native'
