--- src/modules/Home/Home.routes.ts
@@ -1,11 +1,0 @@
-export enum HomeNavigationRoutes {
-  Home = 'Home',
-  Profile = 'Profile',
-  DigitalCv = 'DigitalCv',
-  About = 'About',
-  Experience = 'Experience',
-  Education = 'Education',
-  Skills = 'Skills',
-  MyCourses = 'MyCourses',
-  MyChallenges = 'MyChallenges',
-}
