+++ src/modules/Navigation/Navigation.actions.ts
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
-import { StdObj } from './../../types/general.types'
+import { StdObj } from '../../types/general.types'
 
 export const navigationRef: React.RefObject<any> = React.createRef()
