+++ src/modules/Navigation/Navigation.selector.test.ts
@@ -1,26 +1,23 @@
-import { INITIAL_STATE } from 'modules/Auth/Auth.reducer'
+import { rootStateFixture } from '../../redux/redux.test.fixtures'
+import * as SUT from './Navigation.selector'
 
-import * as SUT from './AppNavigation.selector'
-
-describe('modules/AppNavigation/AppNavigation.selector', () => {
+describe('modules/Navigation/Navigation.selector', () => {
   describe('selector ', () => {
     it('should return expected prop values from initial states', () => {
-      const state = {
-        auth: INITIAL_STATE,
-      }
+      const state = rootStateFixture()
       // when ... we call the selector
       const result = SUT.default(state)
       // then ... should return result as expected
       expect(result).toEqual({ isAuthorised: false })
     })
     it('should select props as expected when available in state', () => {
-      const state = {
+      const state = rootStateFixture({
         auth: {
           refreshToken: 'REFRESH_TOKEN',
           token: 'USER_TOKEN',
           expiresAt: 'EXPIRY_DATE',
         },
-      }
+      })
       // when ... we call the selector
       const result = SUT.default(state)
       // then ... should return result as expected
