+++ src/modules/Navigation/index.ts
@@ -1,0 +1,3 @@
+import Navigation from './Navigation.container'
+
+export default Navigation
