+++ src/modules/RegisterWithEmail/RegisterWithEmail.tsx
@@ -1,9 +1,8 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { YellowCircleRight } from 'assets/images'
 import { Card, LargeHeader, ViewContainer } from 'components'
-import { AuthNavigationRoutes } from 'modules/AppNavigation/Authentication/Authentication.routes'
-import { AuthNavigatorParamsList } from 'modules/AppNavigation/Authentication/Authentication.types'
 import { AuthRegistration } from 'modules/Auth/Auth.types'
+import { AuthNavigationRoutes, AuthNavigatorParamsList } from 'modules/AuthNavigation/AuthNavigation.types'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ScrollView } from 'react-native'
